# pkg/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Ride-Hailing API
  description: |
    Comprehensive ride-hailing platform API supporting riders, drivers, and administrators.

    ## Authentication
    All protected endpoints require a Bearer JWT token in the Authorization header.
    Obtain tokens via the `/api/v1/auth/login` or `/api/v1/auth/register` endpoints.

    ## Rate Limiting
    All endpoints are rate limited. Check `X-RateLimit-Limit`, `X-RateLimit-Remaining`,
    and `X-RateLimit-Reset` response headers.

    ## Idempotency
    For POST/PUT/PATCH requests, include an `Idempotency-Key` header to prevent duplicate processing.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Auth Service
  - url: http://localhost:8081
    description: Rides Service
  - url: http://localhost:8083
    description: Geo Service
  - url: http://localhost:8084
    description: Payments Service
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
...
    Wallet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        balance:
          type: number
          format: double
        currency:
          type: string
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
...
paths:
...
  /api/v1/wallet:
    get:
      tags: [Payments]
      summary: Get wallet balance
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Wallet info
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Wallet'
  /api/v1/wallet/topup:
    post:
      tags: [Payments]
      summary: Top up wallet balance
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  format: double
                  minimum: 1
      responses:
        '200':
          description: Wallet topped up
  /api/v1/wallet/transactions:
    get:
      tags: [Payments]
      summary: Get wallet transaction history
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: Transaction list
  /api/v1/payments/process:
    post:
      tags: [Payments]
      summary: Process payment for a ride
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ride_id, amount, payment_method]
              properties:
                ride_id:
                  type: string
                  format: uuid
                amount:
                  type: number
                  format: double
                payment_method:
                  type: string
                  enum: [wallet, card]
                currency:
                  type: string
                  default: USD
      responses:
        '200':
          description: Payment processed
        '402':
          description: Insufficient funds
  /api/v1/payments/{id}:
    get:
      tags: [Payments]
      summary: Get payment details
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment details
  /api/v1/payments/{id}/refund:
    post:
      tags: [Payments]
      summary: Request payment refund
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Refund processed

  # ==================== GEO ====================
...